fix: 🔧 NyxConfig.fromMap() 统一映射 + ProxyUtils 移除 HandOff 依赖
D — ProxyUtils.fromJvmArgs() 不再调用 HandOff.getConfig()，直接使用 LocateYamlService
E — NyxConfig.fromMap(Map) 替代三处重复的 Map→NyxConfig 映射
    ConfigLoadingController.loading() 缩短为一行 NyxConfig.fromMap(yamlService.load())
