+++ src/main/java/com/nyx/bot/controller/config/ConfigLoadingController.java
@@ -1,6 +1,5 @@
 package com.nyx.bot.controller.config;
 
-import com.nyx.bot.common.config.ConfigConstants;
 import com.nyx.bot.common.config.LocateYamlService;
 import com.nyx.bot.common.core.ApiResponse;
 import com.nyx.bot.common.core.NyxConfig;
@@ -12,8 +11,6 @@ import org.springframework.core.env.MapPropertySource;
 import org.springframework.core.env.PropertySource;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Map;
 
 /**
  * 系统配置
@@ -30,20 +27,7 @@ public class ConfigLoadingController extends BaseController {
 
     @GetMapping
     public ApiResponse<Object> loading() {
-        Map<String, Object> data = yamlService.load();
-        NyxConfig config = new NyxConfig();
-        config.setServerPort(intVal(data, ConfigConstants.SERVER_PORT, 8080));
-        config.setIsServerOrClient(boolVal(data, ConfigConstants.IS_SERVER_OR_CLIENT, true));
-        config.setWsServerUrl(strVal(data, ConfigConstants.WS_SERVER_URL, "/ws/shiro"));
-        config.setWsClientUrl(strVal(data, ConfigConstants.WS_CLIENT_URL, "ws://localhost:3001"));
-        config.setToken(strVal(data, ConfigConstants.TOKEN, ""));
-        config.setHttpProxy(strVal(data, ConfigConstants.HTTP_PROXY, ""));
-        config.setSocksProxy(strVal(data, ConfigConstants.SOCKS_PROXY, ""));
-        config.setProxyUser(strVal(data, ConfigConstants.PROXY_USER, ""));
-        config.setProxyPassword(strVal(data, ConfigConstants.PROXY_PASSWORD, ""));
-        config.setPluginPrefix(boolVal(data, ConfigConstants.PLUGIN_PREFIX, false));
-        config.setPluginName(strVal(data, ConfigConstants.PLUGIN_NAME, ""));
-        return success(config);
+        return success(NyxConfig.fromMap(yamlService.load()));
     }
 
     @PostMapping
@@ -69,24 +53,5 @@ public class ConfigLoadingController extends BaseController {
             }
         }
         return toAjax(true);
-    }
-
-    // ======== 类型安全读取辅助方法 ========
-
-    private static String strVal(Map<String, Object> data, String key, String def) {
-        Object v = data.get(key);
-        return v instanceof String s ? s : def;
-    }
-
-    private static Integer intVal(Map<String, Object> data, String key, Integer def) {
-        Object v = data.get(key);
-        if (v instanceof Number n) return n.intValue();
-        return def;
-    }
-
-    private static Boolean boolVal(Map<String, Object> data, String key, Boolean def) {
-        Object v = data.get(key);
-        if (v instanceof Boolean b) return b;
-        return def;
     }
 }
