+++ src/main/java/com/nyx/bot/utils/http/ProxyUtils.java
@@ -1,8 +1,8 @@
 package com.nyx.bot.utils.http;
 
+import com.nyx.bot.common.config.LocateYamlService;
 import com.nyx.bot.common.core.NyxConfig;
 import com.nyx.bot.common.core.SpringValues;
-import com.nyx.bot.common.config.HandOff;
 import com.nyx.bot.utils.SpringUtils;
 import lombok.extern.slf4j.Slf4j;
 
@@ -81,7 +81,7 @@ public class ProxyUtils {
 
     // region 来源 1：JVM 参数
     private static Proxy fromJvmArgs() {
-        NyxConfig config = HandOff.getConfig();
+        NyxConfig config = NyxConfig.fromMap(new LocateYamlService().load());
         String proxyUrl = config.getHttpProxy() == null || config.getHttpProxy().isEmpty() ? config.getSocksProxy() : config.getHttpProxy();
         String username = config.getProxyUser();
         String password = config.getProxyPassword();
